// src/Api/charge/dto/mipg-api.dto.ts
export class GetMipgChargeDto {
  user?: string;
  ref?: string;
  terminalid?: number;
  status?: number;
  readonly amount?: number;
  payload?: string;
  readonly callbackurl?: string;
  invoiceid?: string;
  userinvoice?: string;
  readonly cardno?: string;
  readonly password?: string;
  readonly nationalcode?: string;
  readonly cardpassword?: string;
  devicetype?: string;
  paytype?: number;
  karmozd?: number;
  wagetype?: number;
  isdirect?: boolean;
}